scraper/pyproject.toml
[project]
name = "scraper"
version = "0.1.0"
description = ""
authors = [
    {name = "Christopher Bonner",email = "cbonner.dev@outlook.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pymysql (>=1.1.1,<2.0.0)",
    "playwright (>=1.50.0,<2.0.0)",
    "bs4 (>=0.0.2,<0.0.3)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "pillow (>=11.1.0,<12.0.0)"
]

[tool.poetry]
packages = [{include = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
